backend/requirements.txt
django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
psycopg2-binary==2.9.1
celery==4.4.7
django-celery-beat==2.0.0
django-celery-results==2.0.1
django-timezone-field==4.2.1
python-crontab==2.6.0
pandas==1.4.3
python-dateutil==2.8.2
numpy==1.22.3
scikit-learn==1.0.2
joblib==1.1.0
tensorflow==2.9.1
docutils==0.19 
Sphinx==5.3.0